# docs/swagger.yaml
definitions:
...
  dto.CreateCommentRequest:
    properties:
      content:
        type: string
      rating:
        maximum: 5
        minimum: 1
        type: number
    required:
      - content
      - rating
    type: object
...
  dto.CreateReplyRequest:
    properties:
      content:
        type: string
      parent_id:
        type: string
    required:
      - content
    type: object
...
  dto.GetCommentsResponse:
    properties:
      book_id:
        type: string
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      rating:
        type: number
      updated_at:
        type: string
      user:
        $ref: "#/definitions/entity.User"
    type: object
  dto.GetRepliesResponse:
    properties:
      comment_id:
        type: string
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      parent_id:
        type: string
      user:
        $ref: "#/definitions/entity.User"
    type: object
...
  entity.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      id:
        type: string
      name:
        type: string
      password:
        type: string
      role:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  models.JSONResponseModel:
    properties:
      data: {}
      errors: {}
      message:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/books/{bookId}/comments:
    get:
      consumes:
        - application/json
      description: Get Comments of a Book
      parameters:
        - description: Book ID
          in: path
          name: bookId
          required: true
          type: string
        - description: Limit
          in: query
          name: limit
          type: integer
        - description: Page
          in: query
          name: page
          type: integer
        - description: Sort By
          in: query
          name: sort_by
          type: string
        - description: Sort Order
          in: query
          name: sort_order
          type: string
        - description: Top Comment Id
          in: query
          name: top_comment_id
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: "#/definitions/models.JSONResponseModel"
              - properties:
                  data:
                    items:
                      allOf:
                        - $ref: "#/definitions/dto.GetCommentsResponse"
                        - properties:
                            book:
                              type: object
                          type: object
                    type: array
                  errors:
                    type: object
                type: object
        "400":
          description: Bad Request
          schema:
            allOf:
              - $ref: "#/definitions/models.JSONResponseModel"
              - properties:
                  data:
                    type: object
                  errors:
                    type: object
                type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
              - $ref: "#/definitions/models.JSONResponseModel"
              - properties:
                  data:
                    type: object
                  errors:
                    type: object
                type: object
      security:
        - BearerAuth: []
      summary: Get Comments of a Book
      tags:
        - Comments
    post:
      consumes:
        - application/json
      description: Create Comments
      parameters:
        - description: Book ID
          in: path
          name: bookId
          required: true
          type: string
        - description: Create Comment Request
          in: body
          name: req
          required: true
          schema:
            $ref: "#/definitions/dto.CreateCommentRequest"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: "#/definitions/models.JSONResponseModel"
              - properties:
                  data:
                    type: object
                  errors:
                    type: object
                type: object
        "400":
          description: Bad Request
          schema:
            allOf:
              - $ref: "#/definitions/models.JSONResponseModel"
              - properties:
                  data:
                    type: object
                  errors:
                    type: object
                type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
              - $ref: "#/definitions/models.JSONResponseModel"
              - properties:
                  data:
                    type: object
                  errors:
                    type: object
                type: object
      security:
        - BearerAuth: []
      summary: Create Comments
      tags:
        - Comments
  /api/books/{bookId}/comments/{comentId}:
    delete:
      consumes:
        - application/json
      description: Delete Comment
      parameters:
        - description: Book ID
          in: path
          name: bookId
          required: true
          type: string
        - description: Comment ID
          in: path
          name: commentId
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: "#/definitions/models.JSONResponseModel"
              - properties:
                  data:
                    type: object
                  errors:
                    type: object
                type: object
        "400":
          description: Bad Request
          schema:
            allOf:
              - $ref: "#/definitions/models.JSONResponseModel"
              - properties:
                  data:
                    type: object
                  errors:
                    type: object
                type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
              - $ref: "#/definitions/models.JSONResponseModel"
              - properties:
                  data:
                    type: object
                  errors:
                    type: object
                type: object
      security:
        - BearerAuth: []
      summary: Delete Comment
      tags:
        - Comments
  /api/comments:
    get:
      consumes:
        - application/json
      description: Get All Comments
      parameters:
        - description: Limit
          in: query
          name: limit
          type: integer
        - description: Page
          in: query
          name: page
          type: integer
        - description: Sort By
          in: query
          name: sort_by
          type: string
        - description: Sort Order
          in: query
          name: sort_order
          type: string
        - description: Top Comment Id
          in: query
          name: top_comment_id
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: "#/definitions/models.JSONResponseModel"
              - properties:
                  data:
                    items:
                      allOf:
                        - $ref: "#/definitions/dto.GetCommentsResponse"
                        - properties:
                            book:
                              type: object
                          type: object
                    type: array
                  errors:
                    type: object
                type: object
        "400":
          description: Bad Request
          schema:
            allOf:
              - $ref: "#/definitions/models.JSONResponseModel"
              - properties:
                  data:
                    type: object
                  errors:
                    type: object
                type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
              - $ref: "#/definitions/models.JSONResponseModel"
              - properties:
                  data:
                    type: object
                  errors:
                    type: object
                type: object
      security:
        - BearerAuth: []
      summary: Get All Comments
      tags:
        - Comments
  /api/comments/{commentId}/replies:
    get:
      consumes:
        - application/json
      description: Get Replies
      parameters:
        - description: Comment ID
          in: path
          name: commentId
          required: true
          type: string
        - description: Limit
          in: query
          name: limit
          type: integer
        - description: Page
          in: query
          name: page
          type: integer
        - description: Sort Order
          in: query
          name: sort_order
          type: string
        - description: Reply Parent ID
          in: query
          name: parent_id
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: "#/definitions/models.JSONResponseModel"
              - properties:
                  data:
                    items:
                      $ref: "#/definitions/dto.GetRepliesResponse"
                    type: array
                  errors:
                    type: object
                type: object
        "400":
          description: Bad Request
          schema:
            allOf:
              - $ref: "#/definitions/models.JSONResponseModel"
              - properties:
                  data:
                    type: object
                  errors:
                    type: object
                type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
              - $ref: "#/definitions/models.JSONResponseModel"
              - properties:
                  data:
                    type: object
                  errors:
                    type: object
                type: object
      security:
        - BearerAuth: []
      summary: Get Replies
      tags:
        - Replies
    post:
      consumes:
        - application/json
      description: Get Reply of a Comment
      parameters:
        - description: Comment ID
          in: path
          name: commentId
          required: true
          type: string
        - description: Request Reply
          in: body
          name: req
          required: true
          schema:
            $ref: "#/definitions/dto.CreateReplyRequest"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: "#/definitions/models.JSONResponseModel"
              - properties:
                  data:
                    type: object
                  errors:
                    type: object
                type: object
        "400":
          description: Bad Request
          schema:
            allOf:
              - $ref: "#/definitions/models.JSONResponseModel"
              - properties:
                  data:
                    type: object
                  errors:
                    type: object
                type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
              - $ref: "#/definitions/models.JSONResponseModel"
              - properties:
                  data:
                    type: object
                  errors:
                    type: object
                type: object
      security:
        - BearerAuth: []
      summary: Get Reply of a Comment
      tags:
        - Replies
  /api/comments/{commentId}/replies/{replyId}:
    delete:
      consumes:
        - application/json
      description: Delete Reply
      parameters:
        - description: Comment ID
          in: path
          name: commentId
          required: true
          type: string
        - description: Reply ID
          in: path
          name: replyId
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: "#/definitions/models.JSONResponseModel"
              - properties:
                  data:
                    type: object
                  errors:
                    type: object
                type: object
        "400":
          description: Bad Request
          schema:
            allOf:
              - $ref: "#/definitions/models.JSONResponseModel"
              - properties:
                  data:
                    type: object
                  errors:
                    type: object
                type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
              - $ref: "#/definitions/models.JSONResponseModel"
              - properties:
                  data:
                    type: object
                  errors:
                    type: object
                type: object
      security:
        - BearerAuth: []
      summary: Delete Reply
      tags:
        - Replies
...
securityDefinitions:
  BearerAuth:
    description: '"Type ''Bearer TOKEN'' to correctly set the API Key"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"